fix: refactor details & fix vc-vendor
